requirements.txt
javalang
numpy
pyparsing
PyGithub
GitPython
